put plugin in separate folder
