+++ application/configs/application.ini
@@ -116,6 +116,11 @@ resources.doctrine.orm.entityManagers.default.metadataDrivers.0.mappingDirs[]   
 resources.doctrine.orm.entityManagers.default.metadataDrivers.0.annotationReaderClass = "Doctrine\Common\Annotations\AnnotationReader"
 resources.doctrine.orm.entityManagers.default.metadataDrivers.0.annotationReaderCache = default
 
+resources.doctrine.orm.entityManagers.default.metadataDrivers.1.adapterClass          = "Doctrine\ORM\Mapping\Driver\AnnotationDriver"
+resources.doctrine.orm.entityManagers.default.metadataDrivers.1.mappingNamespace      = "Plugin"
+resources.doctrine.orm.entityManagers.default.metadataDrivers.1.mappingDirs[]         = APPLICATION_PATH "/Plugin"
+resources.doctrine.orm.entityManagers.default.metadataDrivers.1.annotationReaderClass = "Doctrine\Common\Annotations\AnnotationReader"
+resources.doctrine.orm.entityManagers.default.metadataDrivers.1.annotationReaderCache = default
